Create new component `TrajectoryDrawer` holding all deckgl and websocket logic for one stream
